pyproject.toml
[project]
name = "rascode"
version = "0.1.0"
description = "Raspberry Pi OLED-LCD-HAT-A 三联屏控制与 MCP/Agent 集成（主屏 2 寸 LCD + 双 0.96 寸 OLED）。"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Zhiwei Li", email = "example@example.com" },
]
license = { text = "MIT" }

dependencies = [
  "numpy",
  "pillow",
  "loguru",
  "psutil",
  "luma.oled",
  "luma.lcd",
  "fastmcp",
  "pytest",
  "pytest-cov",
  "ruff",
]

[project.optional-dependencies]
dev = [
  "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
  "pytest",
  "pytest-cov",
  "ruff",
  "mypy",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B"]
ignore = []

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = ["tests"]

